Added Documentation and commented the PostGre stuff and Uncommented H2 related stuff

+++ samples/core/src/main/resources/db/migration/h2/V0001__R001_Create_schema.sql
@@ -30,7 +30,7 @@ CREATE TABLE PRODUCT(
 );
 ALTER TABLE PRODUCT ADD CONSTRAINT PK_PRODUCT PRIMARY KEY(id);
 CREATE TABLE PRODUCT_AUD(
-    revtype smallint,
+    revtype TINYINT,
     description VARCHAR(255),
     name VARCHAR(255),
     pictureId BIGINT,
